Done: Added function ui.UI.cleanupSequence() - removes all players from the board and clears the PlayerPanelArea. It's currently being called from NotifyGameWinner. For that issue - i've thrown the line playerPanelArea.removeAll() to a function of its own (MainWindow.clearPlayerPanelArea()) for convinience, as now we use the function in 2 places (second place being when we add a new playerPanel).
I dont see BOOM and bam anymore - guess you removed them.

To disable throwing dice when on parking we're now sending the player as well the square to 
setSquarePanelContent (In PlayerPanel class), and treating a parking square in a special way.
Still need to check how stuff works for jailed players. Tested for human players.