Document: *IUI *Everything from ui.guiComponents and down.
Reports:
*buyHouse by computer player now works. :)

*Split the if condition of case 0 in the state machine. (line 47 in Monopoly.java)
Now if the square isnt a jail square we dont check the rest of the conditions, so we dont have to count on evaluation order.

ToDo list:
*Fix forfeit action (not removing player) - Might be not having removePlayerIcon after remove player.

*Fix parking (double notification) - NOTE : I changed the condition to player.getParkedOnRound()+2>GameManager.currentGame.getRoundNumber() - Now if because of some unknown glitch a round is skipped or some such thing - the player wont stay stuck in parking for all eternity.
our problem is :
Line 66 of Monopoly.java - case 1 of the state machine.
Line 128 of PlayerPanel.java - disabling the "Throw" button and enabling the end turn button.
Both of these call the method ShouldPlayerMove which prints the line.
If we can change line 66 to do the check otherwise - say, checking the status of the "throw" button (because other places set it correctly), 

then maybe we can fix it.
If not - I don't think a double notification is SO bad.